+++ frontend-react/app/contacts/page.tsx
@@ -2,11 +2,7 @@
 
 import { useState, useEffect } from 'react'
 import { useMediaQuery } from 'react-responsive'
-import HeaderDesktop from '@/components/desktop/layout/HeaderDesktop/HeaderDesktop'
-import FooterDesktop from '@/components/desktop/layout/FooterDesktop/FooterDesktop'
-import HeaderMobi from '@/components/mobi/pageMobi/HomePageMobi/components/HeaderMobi/HeaderMainMobi/ItemHeaderMobi/HeaderMobi'
-import FooterMobi from '@/components/mobi/layout/FooterMobi/FooterMobi'
-import ContactsPageDesktop from '@/components/desktop/pageDesktop/ContactsPageDesktop/ContactsPageDesktop'
+import ContactsPageDesktop from '@/components/desktop/pageDesktop/ContactsPageDesctop/ContactsPageDesktop'
 import ContactsPageMobi from '@/components/mobi/pageMobi/ContactsPageMobi/ContactsPageMobi'
 
 export default function Contacts() {
@@ -27,20 +23,11 @@ export default function Contacts() {
         <div>
             {isDesktop ? (
                 <>
-                    <HeaderDesktop />
-                    <main className="bg-[#101030] text-white">
-                        <ContactsPageDesktop />
-                    </main>
-                    <FooterDesktop />
+                    <ContactsPageDesktop />
                 </>
             ) : (
                 <>
-                    <div className="h-[20px] bg-[#101030]"> </div>
-                    <HeaderMobi />
-                    <main className="bg-[#101030] text-white">
-                        <ContactsPageMobi />
-                    </main>
-                    <FooterMobi />
+                    <ContactsPageMobi />
                 </>
             )}
         </div>
