+++ frontend-react/app/page.tsx
@@ -1,10 +1,7 @@
 'use client'
 import { useState, useEffect } from 'react'
 import { useMediaQuery } from 'react-responsive'
-import HeaderDesktop from '@/components/desktop/layout/HeaderDesktop/HeaderDesktop'
-import FooterDesktop from '@/components/desktop/layout/FooterDesktop/FooterDesktop'
 import HomePageDesktop from '@/components/desktop/pageDesktop/HomePageDesktop/HomePageDesktop'
-import FooterMobi from '@/components/mobi/layout/FooterMobi/FooterMobi'
 import HomePageMobi from '@/components/mobi/pageMobi/HomePageMobi/HomePageMobi'
 import ModalCookieDesktop from '@/components/desktop/layout/ModalDesktop/ModalCookieDesktop'
 
@@ -37,19 +34,12 @@ export default function Home() {
         <div>
             {isDesktop ? (
                 <>
-                    <HeaderDesktop />
-                    <main className="bg-[url('/background/main.webp')] bg-cover bg-no-repeat">
-                        <HomePageDesktop />
-                    </main>
-                    <FooterDesktop />
+                    <HomePageDesktop />
                     {showCookies && <ModalCookieDesktop onClose={handleCloseCookies} />}
                 </>
             ) : (
                 <>
-                    <main className="bg-[#101030]">
-                        <HomePageMobi />
-                    </main>
-                    <FooterMobi />
+                    <HomePageMobi />
                 </>
             )}
         </div>
