+++ frontend-react/app/personalaccount/page.tsx
@@ -1,9 +1,6 @@
 'use client'
 import { useState, useEffect } from 'react'
 import { useMediaQuery } from 'react-responsive'
-import HeaderDesktop from '@/components/desktop/layout/HeaderDesktop/HeaderDesktop'
-import FooterDesktop from '@/components/desktop/layout/FooterDesktop/FooterDesktop'
-import HeaderMobi from '@/components/mobi/pageMobi/HomePageMobi/components/HeaderMobi/HeaderMainMobi/ItemHeaderMobi/HeaderMobi'
 import PersonalAccountDesktop from '@/components/desktop/pageDesktop/PesronalAccountPageDesktop/PersonalAccountPageDesktop'
 import PersonalAccountMobi from '@/components/mobi/pageMobi/PersonalAccountPageMobi/PersonalAccountMobi'
 
@@ -25,15 +22,10 @@ export default function PersonalAccount() {
         <div>
             {isDesktop ? (
                 <>
-                    <HeaderDesktop />
-                    <main className="min-h-[60vh] grow bg-[#101030]">
-                        <PersonalAccountDesktop />
-                    </main>
-                    <FooterDesktop />
+                    <PersonalAccountDesktop />
                 </>
             ) : (
                 <>
-                    <HeaderMobi />
                     <main className="min-h-screen grow bg-[#101030]">
                         <PersonalAccountMobi />
                     </main>
