--- frontend-react/components/desktop/pageDesktop/ContactsPageDesktop/ContactsPageDesktop.tsx
@@ -1,14 +1,0 @@
-import React from 'react'
-import ContactsDesktop from '../ContactsPageDesctop/ContactsDesktop'
-
-const ContactsPageDesktop: React.FC = () => {
-    return (
-        <>
-            <div className="container">
-                <ContactsDesktop />
-            </div>
-        </>
-    )
-}
-
-export default ContactsPageDesktop
