+++ frontend-react/components/desktop/pageDesktop/EventsPageDesktop/data/content.tsx
