+++ frontend-react/components/desktop/pageDesktop/HomePageDesktop/components/EventsSectionDesktop/EventsSectionDesktop.tsx
@@ -1,7 +1,7 @@
 import React, { useRef, useState } from 'react'
 import TitleDesktop from '@/components/desktop/shared/TitleDesktop'
-import ItemEventsDesktop from '@/components/desktop/pageDesktop/HomePageDesktop/components/EventsDesktop/ItemEventsDesktop/ItemEventsDesktop'
-import { contentEventsSection } from '@/components/desktop/pageDesktop/HomePageDesktop/components/EventsDesktop/data/content'
+import ItemEventsDesktop from './ItemEventsDesktop/ItemEventsDesktop'
+import { contentEventsSection } from './data/content'
 const EventsSectionDesktop: React.FC = () => {
     const contentRef = useRef<HTMLDivElement>(null)
     const scrollbarRef = useRef<HTMLDivElement>(null)
