+++ frontend-react/components/mobi/FooterMobi/FooterMobi.tsx
@@ -4,7 +4,7 @@ import React from 'react'
 import Link from 'next/link'
 import Image from 'next/image'
 import { LogoIconMobi, TelegramIconMobi, InstagramIconMobi, LinkedInIconMobi } from '@/components/assets/iconsMobi'
-import { contentFooterMobi, contentFooterMobiImages } from './content'
+import { contentFooterMobi, contentFooterMobiImages } from './data/content'
 
 const FooterMobi: React.FC = () => {
     return (
