+++ frontend-react/components/mobi/pageMobi/HomePageMobi/HomePageMobi.tsx
@@ -1,15 +1,18 @@
 import React from 'react'
-import HeaderMainMobi from './components/HeaderMobi/HeaderMainMobi/HeaderMainMobi'
-import HeaderCardsMobi from './components/HeaderMobi/HeaderCardsMobi/HeaderCardsMobi'
+import HeaderMainMobi from './components/HeaderMainMobi'
+import HeaderCardsMobi from './components/HeaderCardsMobi/HeaderCardsMobi'
 import ProfessionsSectionMobi from './components/ProfessionsSectionMobi/ProfessionsSectionMobi'
 import HowWeWorkMobi from './components/HowWeWorkMobi/HowWeWorkMobi'
 import PromoMobi from './components/PromoMobi/PromoMobi'
 import EventsSectionMobi from './components/EventsSectionMobi/EventsSectionMobi'
+import HeaderMobi from '@/components/mobi/HeaderMobi/HeaderMobi'
+import FooterMobi from '@/components/mobi/FooterMobi/FooterMobi'
 
 const HomePageMobi: React.FC = () => {
     return (
         <>
-            <div>
+            <HeaderMobi />
+            <div className="bg-[#101030]">
                 <HeaderMainMobi />
                 <HeaderCardsMobi />
             </div>
@@ -17,6 +20,7 @@ const HomePageMobi: React.FC = () => {
             <HowWeWorkMobi />
             <PromoMobi />
             <EventsSectionMobi />
+            <FooterMobi />
         </>
     )
 }
