+++ frontend-react/components/mobi/pageMobi/PersonalAccountPageMobi/PersonalAccountMobi.tsx
@@ -1,8 +1,10 @@
 import React, { useState } from 'react'
 import Modal from '@/components/ui/modal'
 import { ChevronLeft } from 'lucide-react'
-import RequiringPaymentMobi from '../../layout/RequiringPaymentMobi/RequiringPaymentMobi'
-import OrderPaymentMobi from '../../layout/OrderPaymentMobi/OrderPaymentMobi'
+import RequiringPaymentMobi from '@/components/mobi/pageMobi/PersonalAccountPageMobi/components/RequiringPaymentMobi/RequiringPaymentMobi'
+import OrderPaymentMobi from '@/components/mobi/pageMobi/PersonalAccountPageMobi/components/OrderPaymentMobi/OrderPaymentMobi'
+import FooterMobi from '@/components/mobi/FooterMobi/FooterMobi'
+import HeaderMobi from '@/components/mobi/HeaderMobi/HeaderMobi'
 
 const PersonalAccountMobi: React.FC = () => {
     const [menuOpen, setMenuOpen] = useState(false)
@@ -29,41 +31,48 @@ const PersonalAccountMobi: React.FC = () => {
     }
 
     return (
-        <div className="flex flex-col items-center justify-center bg-[#101030] px-[15px] pt-[40px]">
-            <h2 className="title28px_mobi_custom mb-4">ЛИЧНЫЙ КАБИНЕТ</h2>
+        <>
+            <HeaderMobi />
+            <div className="min-h-screen grow bg-[#101030]">
+                <div className="flex flex-col items-center justify-center bg-[#101030] px-[15px] pt-[40px]">
+                    <h2 className="title28px_mobi_custom mb-4">ЛИЧНЫЙ КАБИНЕТ</h2>
 
-            <div
-                className="text-text22px_mobi hover:bg-sub-title-gradient-mobi relative 
+                    <div
+                        className="text-text22px_mobi hover:bg-sub-title-gradient-mobi relative 
                 cursor-pointer text-[#878797] hover:bg-clip-text hover:text-transparent "
-            >
-                <button
-                    className="bg-sub-title-gradient-mobi flex items-center bg-clip-text text-transparent underline decoration-[#6C41F3] decoration-1 underline-offset-4"
-                    onClick={() => setMenuOpen(true)}
-                >
-                    {selectedOption}
-                    <div className="-rotate-90">
-                        <ChevronLeft size={30} color="#8333F3" />
+                    >
+                        <button
+                            className="bg-sub-title-gradient-mobi flex items-center bg-clip-text text-transparent underline decoration-[#6C41F3] decoration-1 underline-offset-4"
+                            onClick={() => setMenuOpen(true)}
+                        >
+                            {selectedOption}
+                            <div className="-rotate-90">
+                                <ChevronLeft size={30} color="#8333F3" />
+                            </div>
+                        </button>
                     </div>
-                </button>
-            </div>
 
-            <p className="text-text22px_mobi text-center text-[#353652]">{getContentText()}</p>
-            {menuOpen && (
-                <Modal onClose={() => setMenuOpen(false)} size="small" showCloseButton={false}>
-                    <div className="z-50 flex flex-col space-y-4 py-[36px] text-center">
-                        {options.map((option) => (
-                            <button
-                                key={option}
-                                onClick={() => handleSelectOption(option)}
-                                className="hover:bg-gradient-desktop text18px_mobi cursor-pointer  text-[#878797] hover:bg-clip-text hover:text-transparent hover:underline"
-                            >
-                                {option}
-                            </button>
-                        ))}
-                    </div>
-                </Modal>
-            )}
-        </div>
+                    <p className="text-text22px_mobi text-center text-[#353652]">{getContentText()}</p>
+                    {menuOpen && (
+                        <Modal onClose={() => setMenuOpen(false)} size="small" showCloseButton={false}>
+                            <div className="z-50 flex flex-col space-y-4 py-[36px] text-center">
+                                {options.map((option) => (
+                                    <button
+                                        key={option}
+                                        onClick={() => handleSelectOption(option)}
+                                        className="hover:bg-gradient-desktop text18px_mobi cursor-pointer  text-[#878797] hover:bg-clip-text hover:text-transparent hover:underline"
+                                    >
+                                        {option}
+                                    </button>
+                                ))}
+                            </div>
+                        </Modal>
+                    )}
+                </div>
+                <div />
+            </div>
+            <FooterMobi />
+        </>
     )
 }
